pyproject.toml
[project]
name = "drf-todo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-stubs>=5.2.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
]
